requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.5
sqlmodel==0.0.21
sqlalchemy==2.0.32
pydantic==2.9.2
httpx==0.27.2
openai==1.40.6
python-dotenv==1.0.1
pytest==8.3.2
anyio==4.5.0
